pyproject.toml
[tool.poetry]
name = "djangogames"
version = "0.1.0"
description = "This is a personnel project"
authors = ["Sunny Moses Parlapelly <sunnymoses.parlapelly@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.1"
djangorestframework = "^3.14.0"
whitenoise = "^6.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
